+++ src/Entity/Basket.php
@@ -9,13 +9,14 @@
 
 namespace c975L\ShopBundle\Entity;
 
-use c975L\ShopBundle\Repository\CartRepository;
+use c975L\ShopBundle\Repository\BasketRepository;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
 
-#[ORM\Entity(repositoryClass: CartRepository::class)]
-#[ORM\Table(name: 'shop_cart')]
-class Cart
+#[ORM\Entity(repositoryClass: BasketRepository::class)]
+#[ORM\Table(name: 'shop_basket')]
+class Basket
 {
     #[ORM\Id]
     #[ORM\GeneratedValue]
@@ -23,27 +24,43 @@ class Cart
     private ?int $id = null;
 
     #[ORM\Column(length: 40)]
+    #[Assert\NotBlank]
+    #[Assert\Length(min:40, max: 40)]
     private ?string $identifiant = null;
 
     #[ORM\Column]
     private array $products = [];
 
     #[ORM\Column(length: 50)]
+    #[Assert\NotBlank]
     private ?string $status = null;
 
     #[ORM\Column(length: 100, nullable: true)]
     private ?string $email = null;
 
+    #[ORM\Column(length: 100, nullable: true)]
+    private ?string $address = null;
+
+    #[ORM\Column(length: 100, nullable: true)]
+    private ?string $city = null;
+
+    #[ORM\Column(length: 10, nullable: true)]
+    private ?string $zip = null;
+
     #[ORM\Column(length: 50, nullable: true)]
-    private ?array $address = null;
+    private ?string $country = null;
 
     #[ORM\Column]
+    #[Assert\PositiveOrZero]
     private ?int $total = null;
 
     #[ORM\Column]
+    #[Assert\PositiveOrZero]
     private ?int $quantity = null;
 
     #[ORM\Column(length: 5)]
+    #[Assert\NotBlank]
+    #[Assert\Length(min:3, max: 5)]
     private ?string $currency = null;
 
     #[ORM\Column]
@@ -108,14 +125,50 @@ class Cart
         return $this;
     }
 
-    public function getAddress(): ?array
+    public function getAddress(): ?string
     {
         return $this->address;
     }
 
-    public function setAddress(?array $address): static
+    public function setAddress(?string $address): static
     {
         $this->address = $address;
+
+        return $this;
+    }
+
+    public function getCity(): ?string
+    {
+        return $this->city;
+    }
+
+    public function setCity(?string $city): static
+    {
+        $this->city = $city;
+
+        return $this;
+    }
+
+    public function getZip(): ?string
+    {
+        return $this->zip;
+    }
+
+    public function setZip(?string $zip): static
+    {
+        $this->zip = $zip;
+
+        return $this;
+    }
+
+    public function getCountry(): ?string
+    {
+        return $this->country;
+    }
+
+    public function setCountry(?string $country): static
+    {
+        $this->country = $country;
 
         return $this;
     }